requirements.txt
fastapi
uvicorn
httpx
python-dotenv
truststore
openai
sentence-transformers 